+++ main/src/ca/uwaterloo/flix/language/phase/unification/CaseSetUnification.scala
@@ -16,9 +16,9 @@
 package ca.uwaterloo.flix.language.phase.unification
 
 import ca.uwaterloo.flix.api.Flix
-import ca.uwaterloo.flix.language.ast.Type.eraseAliases
 import ca.uwaterloo.flix.language.ast._
 import ca.uwaterloo.flix.util.Result.{Err, Ok}
+import ca.uwaterloo.flix.util.collection.Bimap
 import ca.uwaterloo.flix.util.{InternalCompilerException, Result}
 
 import scala.annotation.tailrec
@@ -36,64 +36,52 @@ object CaseSetUnification {
       return Ok(Substitution.empty)
     }
 
-    // TODO RESTR-VARS this doesn't help lol
-    //    (tpe1, tpe2) match {
-    //      case (Type.Var(x, _), Type.Var(y, _)) =>
-    //        if (renv.isFlexible(x)) {
-    //          return Ok(Substitution.singleton(x, tpe2)) // 135 hits
-    //        }
-    //        if (renv.isFlexible(y)) {
-    //          return Ok(Substitution.singleton(y, tpe1)) // 0 hits
-    //        }
-    //        if (x == y) {
-    //          return Ok(Substitution.empty) // 0 hits
-    //        }
-    //
-    //      case (Type.Cst(TypeConstructor.CaseAll(_), _), Type.Cst(TypeConstructor.CaseAll(_), _)) =>
-    //        return Ok(Substitution.empty) // 0 hits
-    //
-    //      case (Type.Var(x, _), t2@Type.Cst(tc, _)) if renv.isFlexible(x) => tc match {
-    //        case TypeConstructor.CaseAll(_) =>
-    //          return Ok(Substitution.singleton(x, t2)) // 0 hits
-    //        case TypeConstructor.CaseEmpty(sym) =>
-    //          return Ok(Substitution.singleton(x, t2)) // 0 hits
-    //        case _ => // nop
-    //      }
-    //
-    //      case (Type.Cst(TypeConstructor.CaseEmpty(_), _), Type.Cst(TypeConstructor.CaseEmpty(_), _)) =>
-    //        return Ok(Substitution.empty) //  0 hits
-    //
-    //      case _ => // nop
-    //    }
+    ///
+    /// Get rid of of trivial variable cases.
+    ///
+    (tpe1, tpe2) match {
+      case (t1@Type.Var(x, _), t2) if renv.isFlexible(x) && !t2.typeVars.contains(t1) =>
+        return Ok(Substitution.singleton(x, t2))
+
+      case (t1, t2@Type.Var(x, _))  if renv.isFlexible(x) && !t1.typeVars.contains(t2) =>
+        return Ok(Substitution.singleton(x, t1))
+
+      case _ => // nop
+    }
 
     ///
     /// Run the expensive boolean unification algorithm.
     ///
-    implicit val univ: Universe = Universe(cases, enumSym)
-    val input1 = TypeMinimization.minimizeType(simplify(eraseAliases(tpe1)))
-    val input2 = TypeMinimization.minimizeType(simplify(eraseAliases(tpe2)))
-    booleanUnification(input1, input2, renv)
+
+    val (env, univ) = SetFormula.mkEnv(List(tpe1, tpe2), cases)
+    val input1 = SetFormula.fromCaseType(tpe1, env, univ)
+    val input2 = SetFormula.fromCaseType(tpe2, env, univ)
+
+    booleanUnification(input1, input2, Set.empty, univ, enumSym, env).map {
+      case subst => subst.toTypeSubstitution(enumSym, env)
+    }
   }
 
   /**
     * Returns the most general unifier of the two given set formulas `tpe1` and `tpe2`.
     */
-  private def booleanUnification(tpe1: Type, tpe2: Type, renv: RigidityEnv)(implicit univ: Universe, flix: Flix): Result[Substitution, UnificationError] = {
+  private def booleanUnification(tpe1: SetFormula, tpe2: SetFormula, renv: Set[Int], univ: Set[Int], sym: Symbol.RestrictableEnumSym, env: Bimap[SetFormula.VarOrCase, Int])(implicit flix: Flix): Result[CaseSetSubstitution, UnificationError] = {
     // The boolean expression we want to show is 0.
-    val query = mkEq(tpe1, tpe2)
+    val query = mkEq(tpe1, tpe2, univ)
 
     // Compute the variables in the query.
-    val typeVars = query.typeVars.toList
+    val typeVars = query.freeVars.toList
 
     // Compute the flexible variables.
-    val flexibleTypeVars = renv.getFlexibleVarsOf(typeVars)
+    val flexibleTypeVars = typeVars.filterNot(renv.contains)
 
     // Determine the order in which to eliminate the variables.
-    val freeVars = computeVariableOrder(flexibleTypeVars)
+    //    val freeVars = computeVariableOrder(flexibleTypeVars)
+    val freeVars = flexibleTypeVars
 
     // Eliminate all variables.
     try {
-      val subst = successiveVariableElimination(query, freeVars)
+      val subst = successiveVariableElimination(query, freeVars, univ)
 
       //    if (!subst.isEmpty) {
       //      val s = subst.toString
@@ -106,7 +94,10 @@ object CaseSetUnification {
 
       Ok(subst)
     } catch {
-      case SetUnificationException => Err(UnificationError.MismatchedBools(tpe1, tpe2)) // TODO make setty
+      case SetUnificationException =>
+        val t1 = SetFormula.toCaseType(tpe1, sym, env, SourceLocation.Unknown)
+        val t2 = SetFormula.toCaseType(tpe2, sym, env, SourceLocation.Unknown)
+        Err(UnificationError.MismatchedBools(t1, t2)) // TODO make setty
     }
   }
 
@@ -133,21 +124,23 @@ object CaseSetUnification {
   /**
     * Performs successive variable elimination on the given set expression `f`.
     */
-  private def successiveVariableElimination(f: Type, fvs: List[Type.Var])(implicit univ: Universe, flix: Flix): Substitution = fvs match {
+  private def successiveVariableElimination(f: SetFormula, fvs: List[Int], univ: Set[Int])(implicit flix: Flix): CaseSetSubstitution = fvs match {
     case Nil =>
       // Determine if f is necessarily empty when all (rigid) variables and constants are made flexible.
-      if (isEmpty(dnf(f)))
-        Substitution.empty
+      if (eval(f, univ).isEmpty)
+        CaseSetSubstitution.empty
       else
         throw SetUnificationException
 
     case x :: xs =>
-      val t0 = Substitution.singleton(x.sym, mkEmpty(univ.enumSym))(f)
-      val t1 = Substitution.singleton(x.sym, mkAll(univ.enumSym))(f)
-      val se = successiveVariableElimination(mkIntersection(t0, t1, univ.enumSym), xs)
+      val t0 = CaseSetSubstitution.singleton(x, SetFormula.Empty)(f, univ)
+      val t1 = CaseSetSubstitution.singleton(x, SetFormula.Cst(univ))(f, univ)
+      val se = successiveVariableElimination(mkIntersection(t0, t1, univ), xs, univ)
 
-      val f1 = TypeMinimization.minimizeType(mkUnion(se(t0), mkIntersection(x, mkComplement(se(t1), univ.enumSym), univ.enumSym), univ.enumSym))
-      val st = Substitution.singleton(x.sym, f1)
+      val f1 = mkUnion(se(t0, univ), mkIntersection(SetFormula.Var(x), mkComplement(se(t1, univ), univ), univ), univ)
+//      val f2 = SetFormulaAlg.simplifyByExhaustiveEvaluation(f1)(univ)
+      val f2 = minViaTable(f1, univ)
+      val st = CaseSetSubstitution.singleton(x, f2)
       st ++ se
   }
 
@@ -159,31 +152,28 @@ object CaseSetUnification {
   /**
     * To unify two set formulas p and q it suffices to unify t = (p ∧ ¬q) ∨ (¬p ∧ q) and check t = 0.
     */
-  private def mkEq(p: Type, q: Type)(implicit univ: Universe): Type = mkUnion(mkIntersection(p, mkComplement(q, univ.enumSym), univ.enumSym), mkIntersection(mkComplement(p, univ.enumSym), q, univ.enumSym), univ.enumSym)
+  private def mkEq(p: SetFormula, q: SetFormula, univ: Set[Int]): SetFormula = mkUnion(mkIntersection(p, mkComplement(q, univ), univ), mkIntersection(mkComplement(p, univ), q, univ), univ)
 
   /**
     * Returns the negation of the set formula `tpe0`.
     */
   // NB: The order of cases has been determined by code coverage analysis.
-  def mkComplement(tpe0: Type, sym: Symbol.RestrictableEnumSym): Type = tpe0 match {
-    case ALL() =>
-      mkEmpty(sym)
+  def mkComplement(tpe0: SetFormula, univ: Set[Int]): SetFormula = tpe0 match {
+    case SetFormula.Cst(s) =>
+      SetFormula.Cst(univ -- s)
 
-    case EMPTY() =>
-      mkAll(sym)
-
-    case COMPLEMENT(x) =>
+    case SetFormula.Not(x) =>
       x
 
     // ¬(¬x ∨ y) => x ∧ ¬y
-    case UNION(COMPLEMENT(x), y) =>
-      mkIntersection(x, mkComplement(y, sym), sym)
+    case SetFormula.Or(SetFormula.Not(x), y) =>
+      mkIntersection(x, mkComplement(y, univ), univ)
 
     // ¬(x ∨ ¬y) => ¬x ∧ y
-    case UNION(x, COMPLEMENT(y)) =>
-      mkIntersection(mkComplement(x, sym), y, sym)
+    case SetFormula.Or(x, SetFormula.Not(y)) =>
+      mkIntersection(mkComplement(x, univ), y, univ)
 
-    case _ => Type.Apply(Type.Cst(TypeConstructor.CaseComplement(sym), tpe0.loc), tpe0, tpe0.loc)
+    case _ => SetFormula.Not(tpe0)
   }
 
   /**
@@ -191,94 +181,75 @@ object CaseSetUnification {
     */
   // NB: The order of cases has been determined by code coverage analysis.
   @tailrec
-  def mkIntersection(tpe1: Type, tpe2: Type, sym: Symbol.RestrictableEnumSym): Type = (tpe1, tpe2) match {
-    // T ∧ x => x
-    case (ALL(), _) =>
-      tpe2
-
-    // x ∧ T => x
-    case (_, ALL()) =>
-      tpe1
-
-    // F ∧ x => F
-    case (EMPTY(), _) =>
-      mkEmpty(sym)
-
-    // x ∧ F => F
-    case (_, EMPTY()) =>
-      mkEmpty(sym)
-
-    // C ∧ D => F
-    case (CONSTANT(x1), CONSTANT(x2)) if x1 != x2 =>
-      mkEmpty(sym)
-
-    // C ∧ ¬D => C
-    case (x1@CONSTANT(_), COMPLEMENT(x2@CONSTANT(_))) if x1 != x2 =>
-      x1
-
-    // ¬C ∧ D => D
-    case (COMPLEMENT(x1@CONSTANT(_)), x2@CONSTANT(_)) if x1 != x2 =>
+  def mkIntersection(tpe1: SetFormula, tpe2: SetFormula, univ: Set[Int]): SetFormula = (tpe1, tpe2) match {
+    case (SetFormula.Cst(x1), x2) if x1 == univ =>
       x2
 
+    case (x1, SetFormula.Cst(x2)) if x2 == univ =>
+      x1
+
+    case (SetFormula.Cst(x1), SetFormula.Cst(x2)) =>
+      SetFormula.Cst(x1 & x2)
+
     // ¬x ∧ (x ∨ y) => ¬x ∧ y
-    case (COMPLEMENT(x1), UNION(x2, y)) if x1 == x2 =>
-      mkIntersection(mkComplement(x1, sym), y, sym)
+    case (SetFormula.Not(x1), SetFormula.Or(x2, y)) if x1 == x2 =>
+      mkIntersection(mkComplement(x1, univ), y, univ)
 
     // x ∧ ¬x => F
-    case (x1, COMPLEMENT(x2)) if x1 == x2 =>
-      mkEmpty(sym)
+    case (x1, SetFormula.Not(x2)) if x1 == x2 =>
+      SetFormula.Empty
 
     // ¬x ∧ x => F
-    case (COMPLEMENT(x1), x2) if x1 == x2 =>
-      mkEmpty(sym)
+    case (SetFormula.Not(x1), x2) if x1 == x2 =>
+      SetFormula.Empty
 
     // x ∧ (x ∧ y) => (x ∧ y)
-    case (x1, INTERSECTION(x2, y)) if x1 == x2 =>
-      mkIntersection(x1, y, sym)
+    case (x1, SetFormula.And(x2, y)) if x1 == x2 =>
+      mkIntersection(x1, y, univ)
 
     // x ∧ (y ∧ x) => (x ∧ y)
-    case (x1, INTERSECTION(y, x2)) if x1 == x2 =>
-      mkIntersection(x1, y, sym)
+    case (x1, SetFormula.And(y, x2)) if x1 == x2 =>
+      mkIntersection(x1, y, univ)
 
     // (x ∧ y) ∧ x) => (x ∧ y)
-    case (INTERSECTION(x1, y), x2) if x1 == x2 =>
-      mkIntersection(x1, y, sym)
+    case (SetFormula.And(x1, y), x2) if x1 == x2 =>
+      mkIntersection(x1, y, univ)
 
     // (x ∧ y) ∧ y) => (x ∧ y)
-    case (INTERSECTION(x, y1), y2) if y1 == y2 =>
-      mkIntersection(x, y1, sym)
+    case (SetFormula.And(x, y1), y2) if y1 == y2 =>
+      mkIntersection(x, y1, univ)
 
     // x ∧ (x ∨ y) => x
-    case (x1, UNION(x2, _)) if x1 == x2 =>
+    case (x1, SetFormula.Or(x2, _)) if x1 == x2 =>
       x1
 
     // (x ∨ y) ∧ x => x
-    case (UNION(x1, _), x2) if x1 == x2 =>
+    case (SetFormula.Or(x1, _), x2) if x1 == x2 =>
       x1
 
     // x ∧ (y ∧ ¬x) => F
-    case (x1, INTERSECTION(_, COMPLEMENT(x2))) if x1 == x2 =>
-      mkEmpty(sym)
+    case (x1, SetFormula.And(_, SetFormula.Not(x2))) if x1 == x2 =>
+      SetFormula.Empty
 
     // (¬x ∧ y) ∧ x => F
-    case (INTERSECTION(COMPLEMENT(x1), _), x2) if x1 == x2 =>
-      mkEmpty(sym)
+    case (SetFormula.And(SetFormula.Not(x1), _), x2) if x1 == x2 =>
+      SetFormula.Empty
 
     // x ∧ ¬(x ∨ y) => F
-    case (x1, COMPLEMENT(UNION(x2, _))) if x1 == x2 =>
-      mkEmpty(sym)
+    case (x1, SetFormula.Not(SetFormula.Or(x2, _))) if x1 == x2 =>
+      SetFormula.Empty
 
     // ¬(x ∨ y) ∧ x => F
-    case (COMPLEMENT(UNION(x1, _)), x2) if x1 == x2 =>
-      mkEmpty(sym)
+    case (SetFormula.Not(SetFormula.Or(x1, _)), x2) if x1 == x2 =>
+      SetFormula.Empty
 
     // x ∧ (¬x ∧ y) => F
-    case (x1, INTERSECTION(COMPLEMENT(x2), _)) if x1 == x2 =>
-      mkEmpty(sym)
+    case (x1, SetFormula.And(SetFormula.Not(x2), _)) if x1 == x2 =>
+      SetFormula.Empty
 
     // (¬x ∧ y) ∧ x => F
-    case (INTERSECTION(COMPLEMENT(x1), _), x2) if x1 == x2 =>
-      mkEmpty(sym)
+    case (SetFormula.And(SetFormula.Not(x1), _), x2) if x1 == x2 =>
+      SetFormula.Empty
 
     // x ∧ x => x
     case _ if tpe1 == tpe2 => tpe1
@@ -290,7 +261,7 @@ object CaseSetUnification {
       //        println(s.substring(0, Math.min(len, 300)))
       //      }
 
-      Type.Apply(Type.Apply(Type.Cst(TypeConstructor.CaseIntersection(sym), tpe1.loc), tpe1, tpe1.loc), tpe2, tpe1.loc)
+      SetFormula.And(tpe1, tpe2)
   }
 
   /**
@@ -298,48 +269,45 @@ object CaseSetUnification {
     */
   // NB: The order of cases has been determined by code coverage analysis.
   @tailrec
-  def mkUnion(tpe1: Type, tpe2: Type, sym: Symbol.RestrictableEnumSym): Type = (tpe1, tpe2) match {
-    // T ∨ x => T
-    case (ALL(), _) =>
-      mkAll(sym)
+  def mkUnion(tpe1: SetFormula, tpe2: SetFormula, univ: Set[Int]): SetFormula = (tpe1, tpe2) match {
+    case (SetFormula.Cst(x1), x2) if x1 == univ =>
+      SetFormula.Cst(x1)
 
-    // F ∨ y => y
-    case (EMPTY(), _) =>
-      tpe2
+    case (x1, SetFormula.Cst(x2)) if x2 == univ =>
+      SetFormula.Cst(x2)
 
-    // x ∨ F => x
-    case (_, EMPTY()) =>
-      tpe1
+    case (SetFormula.Cst(s1), SetFormula.Cst(s2)) =>
+      SetFormula.Cst(s1 ++ s2)
 
     // x ∨ (y ∨ x) => x ∨ y
-    case (x1, UNION(y, x2)) if x1 == x2 =>
-      mkUnion(x1, y, sym)
+    case (x1, SetFormula.Or(y, x2)) if x1 == x2 =>
+      mkUnion(x1, y, univ)
 
     // (x ∨ y) ∨ x => x ∨ y
-    case (UNION(x1, y), x2) if x1 == x2 =>
-      mkUnion(x1, y, sym)
+    case (SetFormula.Or(x1, y), x2) if x1 == x2 =>
+      mkUnion(x1, y, univ)
 
     // ¬x ∨ x => T
-    case (COMPLEMENT(x), y) if x == y =>
-      mkAll(sym)
+    case (SetFormula.Not(x), y) if x == y =>
+      SetFormula.Cst(univ)
 
     // x ∨ ¬x => T
-    case (x, COMPLEMENT(y)) if x == y =>
-      mkAll(sym)
+    case (x, SetFormula.Not(y)) if x == y =>
+      SetFormula.Cst(univ)
 
     // (¬x ∨ y) ∨ x) => T
-    case (UNION(COMPLEMENT(x), _), y) if x == y =>
-      mkAll(sym)
+    case (SetFormula.Or(SetFormula.Not(x), _), y) if x == y =>
+      SetFormula.Cst(univ)
 
     // x ∨ (¬x ∨ y) => T
-    case (x, UNION(COMPLEMENT(y), _)) if x == y =>
-      mkAll(sym)
+    case (x, SetFormula.Or(SetFormula.Not(y), _)) if x == y =>
+      SetFormula.Cst(univ)
 
     // x ∨ (y ∧ x) => x
-    case (x1, INTERSECTION(_, x2)) if x1 == x2 => x1
+    case (x1, SetFormula.And(_, x2)) if x1 == x2 => x1
 
     // (y ∧ x) ∨ x => x
-    case (INTERSECTION(_, x1), x2) if x1 == x2 => x1
+    case (SetFormula.And(_, x1), x2) if x1 == x2 => x1
 
     // x ∨ x => x
     case _ if tpe1 == tpe2 =>
@@ -353,282 +321,39 @@ object CaseSetUnification {
       //                println(s.substring(0, Math.min(len, 300)))
       //              }
 
-      Type.Apply(Type.Apply(Type.Cst(TypeConstructor.CaseUnion(sym), tpe1.loc), tpe1, tpe1.loc), tpe2, tpe1.loc)
+      SetFormula.Or(tpe1, tpe2)
   }
 
   /**
-    * Returns the difference of the given types.
+    * Evaluates the set formula. Assumes there are no variables in the formula.
     */
-  private def mkDifference(tpe1: Type, tpe2: Type)(implicit univ: Universe): Type = mkIntersection(tpe1, mkComplement(tpe2, univ.enumSym), univ.enumSym)
+  def eval(f: SetFormula, univ: Set[Int]): Set[Int] = f match {
+    case SetFormula.Cst(s) => s
+    case SetFormula.Not(f) => univ -- eval(f, univ)
+    case SetFormula.And(f1, f2) => eval(f1, univ) & eval(f2, univ)
+    case SetFormula.Or(f1, f2) => eval(f1, univ) ++ eval(f2, univ)
+    case SetFormula.Var(x) => throw InternalCompilerException("unexpected var", SourceLocation.Unknown)
+  }
 
-  // TODO RESTR-VARS docs
-  // TODO RESTR-VARS caching
-  private def mkEmpty(sym: Symbol.RestrictableEnumSym): Type = Type.Cst(TypeConstructor.CaseEmpty(sym), SourceLocation.Unknown)
+  private def minViaTable(f: SetFormula, univ: Set[Int]): SetFormula = {
+    val bot = SetFormula.Cst(Set.empty): SetFormula
+    val top = SetFormula.Cst(univ): SetFormula
 
-  // TODO RESTR-VARS docs
-  // TODO RESTR-VARS caching
-  private def mkAll(sym: Symbol.RestrictableEnumSym): Type = Type.Cst(TypeConstructor.CaseAll(sym), SourceLocation.Unknown)
+    def visit(fvs: List[Int]): List[(SetFormula, Map[Int, SetFormula])] = fvs match {
+    case Nil => List((top, Map.empty))
+      case v :: vs =>
+        visit(vs) flatMap {
+          case (p, partialSubst) =>
+            val p1 = mkIntersection(SetFormula.Not(SetFormula.Var(v)), p, univ)
+            val p2 = mkIntersection(SetFormula.Var(v), p, univ)
+            List((p1, partialSubst + (v -> bot)), (p2, partialSubst + (v -> top)))
+        }
+    }
 
-  private object COMPLEMENT {
-    @inline
-    def unapply(tpe: Type): Option[Type] = tpe match {
-      case Type.Apply(Type.Cst(TypeConstructor.CaseComplement(_), _), x, _) => Some(x)
-      case _ => None
+    visit(f.freeVars.toList).foldLeft(bot) {
+      case (acc, (p, subst)) => mkUnion(acc, mkIntersection(p, eval(f, subst, univ), univ), univ)
     }
   }
 
-  private object INTERSECTION {
-    @inline
-    def unapply(tpe: Type): Option[(Type, Type)] = tpe match {
-      case Type.Apply(Type.Apply(Type.Cst(TypeConstructor.CaseIntersection(_), _), x, _), y, _) => Some((x, y))
-      case _ => None
-    }
-  }
-
-  private object UNION {
-    @inline
-    def unapply(tpe: Type): Option[(Type, Type)] = tpe match {
-      case Type.Apply(Type.Apply(Type.Cst(TypeConstructor.CaseUnion(_), _), x, _), y, _) => Some((x, y))
-      case _ => None
-    }
-  }
-
-  private object CONSTANT {
-    @inline
-    def unapply(tpe: Type): Option[Symbol.RestrictableCaseSym] = tpe match {
-      case Type.Cst(TypeConstructor.CaseConstant(sym), _) => Some(sym)
-      case _ => None
-    }
-  }
-
-  private object EMPTY {
-    @inline
-    def unapply(tpe: Type): Boolean = tpe match {
-      case Type.Cst(TypeConstructor.CaseEmpty(_), _) => true
-      case _ => false
-    }
-  }
-
-  private object ALL {
-    @inline
-    def unapply(tpe: Type): Boolean = tpe match {
-      case Type.Cst(TypeConstructor.CaseAll(_), _) => true
-      case _ => false
-    }
-  }
-
-  private object VAR {
-    @inline
-    def unapply(tpe: Type): Option[Symbol.KindedTypeVarSym] = tpe match {
-      case Type.Var(sym, _) => Some(sym)
-      case _ => None
-    }
-  }
-
-  /**
-    * An atom is a constant or a variable.
-    */
-  private sealed trait Atom
-
-  private object Atom {
-    case class Var(sym: Symbol.KindedTypeVarSym) extends Atom
-
-    case class Case(sym: Symbol.RestrictableCaseSym) extends Atom
-  }
-
-  /**
-    * A literal is a negated or un-negated atom.
-    */
-  private sealed trait Literal
-
-  private object Literal {
-    case class Positive(atom: Atom) extends Literal
-
-    case class Negative(atom: Atom) extends Literal
-  }
-
-  /**
-    * A DNF intersection is a set of literals.
-    */
-  private type Intersection = Set[Literal]
-
-  /**
-    * A DNF formula is either:
-    * - a union of intersections of literals.
-    * - the universal set
-    */
-  private sealed trait Dnf
-
-  private object Dnf {
-    case class Union(inters: Set[Intersection]) extends Dnf // TODO should be a list to avoid extra comparisons?
-
-    /**
-      * The bottom value is an empty union.
-      */
-    val Empty = Union(Set())
-
-    /**
-      * The top value is an empty intersection.
-      */
-    val All = Union(Set(Set()))
-  }
-
-  /**
-    * An NNF formula is a formula where all negations are on atoms.
-    */
-  private sealed trait Nnf
-
-  private object Nnf {
-    case class Union(tpe1: Nnf, tpe2: Nnf) extends Nnf
-
-    case class Intersection(tpe1: Nnf, tpe2: Nnf) extends Nnf
-
-    case class Singleton(tpe: Literal) extends Nnf
-
-    case object Empty extends Nnf
-
-    case object All extends Nnf
-  }
-
-  private def simplify(t: Type)(implicit univ: Universe): Type = {
-    fromDnf(dnf(t))
-  }
-
-  private def fromDnf(t: Dnf)(implicit univ: Universe): Type = t match {
-    case Dnf.Union(inters) => inters.filterNot(isEmptyIntersection).map(fromIntersection).reduceOption(mkUnion(_, _, univ.enumSym)).getOrElse(mkEmpty(univ.enumSym))
-  }
-
-  private def fromAtom(a: Atom)(implicit univ: Universe): Type = a match {
-    case Atom.Var(sym) => Type.Var(sym, SourceLocation.Unknown)
-    case Atom.Case(sym) => Type.Cst(TypeConstructor.CaseConstant(sym), SourceLocation.Unknown)
-  }
-
-  private def fromLiteral(l: Literal)(implicit univ: Universe): Type = l match {
-    case Literal.Positive(atom) => fromAtom(atom)
-    case Literal.Negative(atom) => mkComplement(fromAtom(atom), univ.enumSym)
-  }
-
-  private def fromIntersection(i: Intersection)(implicit univ: Universe): Type = {
-    i.map(fromLiteral).reduceOption(mkIntersection(_, _, univ.enumSym)).getOrElse(mkAll(univ.enumSym))
-  }
-
-  /**
-    * Converts the given type to DNF
-    */
-  private def dnf(t: Type)(implicit univ: Universe): Dnf = {
-    val n = nnf(t)
-    val d = nnfToDnf(n)
-    d
-  }
-
-  /**
-    * Converts the given type to NNF.
-    */
-  private def nnf(t: Type)(implicit univ: Universe): Nnf = t match {
-    case CONSTANT(sym) => Nnf.Singleton(Literal.Positive(Atom.Case(sym)))
-    case VAR(sym) => Nnf.Singleton(Literal.Positive(Atom.Var(sym)))
-    case COMPLEMENT(tpe) => nnfNot(tpe)
-    case UNION(tpe1, tpe2) => Nnf.Union(nnf(tpe1), nnf(tpe2))
-    case INTERSECTION(tpe1, tpe2) => Nnf.Intersection(nnf(tpe1), nnf(tpe2))
-    case EMPTY() => Nnf.Empty
-    case ALL() => Nnf.All
-    case _ => throw InternalCompilerException(s"unexpected type: $t", t.loc)
-  }
-
-  /**
-    * Converts the complement of the given type to NNF.
-    */
-  private def nnfNot(t: Type)(implicit univ: Universe): Nnf = t match {
-    case CONSTANT(sym) => Nnf.Singleton(Literal.Negative(Atom.Case(sym)))
-    case VAR(sym) => Nnf.Singleton(Literal.Negative(Atom.Var(sym)))
-    case COMPLEMENT(tpe) => nnf(tpe)
-    case UNION(tpe1, tpe2) => Nnf.Intersection(
-      nnf(mkComplement(tpe1, univ.enumSym)),
-      nnf(mkComplement(tpe2, univ.enumSym))
-    )
-    case INTERSECTION(tpe1, tpe2) => Nnf.Union(
-      nnf(mkComplement(tpe1, univ.enumSym)),
-      nnf(mkComplement(tpe2, univ.enumSym))
-    )
-    case EMPTY() => Nnf.All // MATT ???
-    case ALL() => Nnf.Empty // MATT ???
-    case _ => throw InternalCompilerException(s"unexpected type: $t", t.loc)
-  }
-
-  /**
-    * Converts the given type from NNF to DNF.
-    */
-  private def nnfToDnf(t: Nnf): Dnf = t match {
-    case Nnf.Union(tpe1, tpe2) => union(nnfToDnf(tpe1), nnfToDnf(tpe2))
-    case Nnf.Intersection(tpe1, tpe2) => intersect(nnfToDnf(tpe1), nnfToDnf(tpe2))
-    case Nnf.Singleton(tpe) => Dnf.Union(Set(Set(tpe)))
-    case Nnf.Empty => Dnf.Empty
-    case Nnf.All => Dnf.All
-  }
-
-  /**
-    * Calculates the intersection of two DNF sets.
-    */
-  private def intersect(t1: Dnf, t2: Dnf): Dnf = (t1, t2) match {
-    case (Dnf.Union(inters1), Dnf.Union(inters2)) =>
-      val inters = for {
-        inter1 <- inters1
-        inter2 <- inters2
-      } yield inter1 ++ inter2
-      Dnf.Union(inters)
-  }
-
-  /**
-    * Calculates the union of two DNF sets.
-    */
-  private def union(t1: Dnf, t2: Dnf): Dnf = (t1, t2) match {
-    case (Dnf.Union(inters1), Dnf.Union(inters2)) => Dnf.Union(inters1 ++ inters2)
-  }
-
-  /**
-    * Returns true if the given DNF set represents an empty set.
-    */
-  private def isEmpty(t1: Dnf)(implicit univ: Universe): Boolean = t1 match {
-    case Dnf.Union(inters) => inters.forall(isEmptyIntersection)
-  }
-
-  /**
-    * Returns true if `t1` represents an empty intersection of effects.
-    */
-  private def isEmptyIntersection(t1: Intersection)(implicit univ: Universe): Boolean = {
-    val pos = t1.collect {
-      case Literal.Positive(atom) => atom
-    }
-    val neg = t1.collect {
-      case Literal.Negative(atom) => atom
-    }
-
-    // an intersection is empty if any of the following is true:
-
-    // 1. It contains two different positive constants
-    val diffConst = pos.collect {
-      case Atom.Case(sym) => sym
-    }.size > 1
-
-    // 2. It contains an atom in both the positive and negative sets
-    val negation = (pos & neg).nonEmpty
-
-    // 3. It contains all the negative constants
-    val allNegConsts = univ.cases.forall {
-      case c => neg.exists {
-        case Atom.Case(sym) => c == sym
-        case Atom.Var(_) => false
-      }
-    }
-
-    //    if (diffConst || negation || allNegConsts) {
-    //      println(s" IS EMPTY: $t1")
-    //    } else {
-    //      println(s"NOT EMPTY: $t1")
-    //    }
-
-    diffConst || negation || allNegConsts
-  }
-
-  case class Universe(cases: List[Symbol.RestrictableCaseSym], enumSym: Symbol.RestrictableEnumSym)
+  private def eval(f: SetFormula, m: Map[Int, SetFormula], univ: Set[Int]): SetFormula = SetFormula.map(f)(m)(univ)
 }
